project/requirements.txt
absl-py==1.4.0
altgraph==0.17.2
anyio==3.6.2
appdirs==1.4.4
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.20.1
backcall==0.2.0
beautifulsoup4==4.8.2
binaryornot==0.4.4
bleach==5.0.1
bottle==0.12.21
bottle-websocket==0.2.9
cachetools==5.3.0
cassandra-driver==3.25.0
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
compressed-rtf==1.0.6
contourpy==1.0.5
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.1.7
django-cors-headers==3.11.0
djangorestframework==3.12.2
dnspython==2.2.1
docx2txt==0.8
drf-yasg==1.20.0
ebcdic==1.1.1
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0
executing==1.1.1
extract-msg==0.28.7
facebook-sdk==3.1.0
fastapi==0.89.1
fastjsonschema==2.16.2
filelock==3.6.0
flatbuffers==23.1.21
fontawesomefree==6.3.0
fonttools==4.37.4
fst-pso==1.8.1
ftfy==6.1.1
future==0.18.2
Fuzzy==1.2.2
FuzzyTM==2.0.5
gast==0.4.0
gdown==4.7.1
gensim==4.3.0
geomet==0.2.1.post1
gevent==21.12.0
gevent-websocket==0.10.1
gns3fy==0.8.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
idna==3.3
IMAPClient==2.1.0
imgviz==1.7.2
inflection==0.5.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.11.0
kiwisolver==1.4.4
labelme==5.1.1
labelme2yolo==0.0.8
libclang==15.0.6.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
matplotlib-label-lines==0.5.1
MetaTrader5==5.0.43
miniful==0.0.6
mistune==2.0.4
mixed-naive-bayes==0.0.3
more-itertools==9.0.0
mpmath==1.2.1
multitasking==0.0.11
mysql-connector-python==8.0.27
natsort==8.3.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.4.12
numpy==1.23.2
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.7.0.72
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20191110
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
pure-eval==0.2.2
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
pyFUME==0.2.25
Pygments==2.13.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pymongo==4.3.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.3.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
python-pptx==0.6.21
python-slugify==7.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywaffle==1.1.0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.8
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
redis==4.4.2
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
selenium==4.5.0
Send2Trash==1.8.0
sentry-sdk==1.16.0
simpful==2.10.0
six==1.12.0
sklearn==0.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.46
sqlparse==0.4.2
stack-data==0.5.1
starlette==0.22.0
sympy==1.11.1
tabulate==0.8.9
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.16.0
text-unidecode==1.3
textblob==0.17.1
textract==1.6.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.1.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ultralytics==8.0.49
uritemplate==4.1.1
urllib3==1.26.14
utility==1.0
uvicorn==0.20.0
virtualenv==20.14.1
wcwidth==0.2.5
webdriver-manager==3.8.4
webencodings==0.5.1
Werkzeug==2.2.3
whichcraft==0.6.1
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.0.3
yfinance==0.1.86
zope.event==4.5.0
zope.interface==5.4.0
